drivers/kernel_test: rename the kernel test driver
- rename the kernel_test_drv as os_test_drv because kernel_test contains kernel and network TCs and also may contain filesystem TCs.
- remove unnecessary header files
