+++ apps/examples/testcase/le_tc/kernel/tc_group.c
@@ -26,7 +26,7 @@
 #include <tinyara/config.h>
 #include <stdio.h>
 #include <errno.h>
-#include <tinyara/kernel_test_drv.h>
+#include <tinyara/os_api_test_drv.h>
 #include "tc_internal.h"
