+++ apps/examples/testcase/le_tc/kernel/tc_semaphore.c
@@ -31,7 +31,7 @@
 #include <sys/time.h>
 #include <sys/types.h>
 #include <sys/ioctl.h>
-#include <tinyara/kernel_test_drv.h>
+#include <tinyara/os_api_test_drv.h>
 #include "tc_internal.h"
 
 #define PSHARED			0
