+++ apps/examples/testcase/le_tc/network/tc_net_pbuf.c
@@ -28,7 +28,7 @@
 #include <netutils/netlib.h>
 #include <sys/socket.h>
 #include <lwip/pbuf.h>
-#include <tinyara/kernel_test_drv.h>
+#include <tinyara/os_api_test_drv.h>
 #include <sys/ioctl.h>
 #include "tc_internal.h"
