+++ os/drivers/Makefile
@@ -79,7 +79,7 @@ include gpio$(DELIM)Make.defs
 include memory$(DELIM)Make.defs
 include i2c$(DELIM)Make.defs
 include iotdev$(DELIM)Make.defs
-include kernel_test$(DELIM)Make.defs
+include os_api_test$(DELIM)Make.defs
 include lcd$(DELIM)Make.defs
 include lwnl${DELIM}Make.defs
 include net$(DELIM)Make.defs
