--- os/drivers/kernel_test/Make.defs
@@ -1,25 +1,0 @@
-###########################################################################
-#
-# Copyright 2017 Samsung Electronics All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
-# either express or implied. See the License for the specific
-# language governing permissions and limitations under the License.
-#
-###########################################################################
-
-ifeq ($(CONFIG_DRIVERS_KERNEL_TEST),y)
-include kernel_test$(DELIM)kernel$(DELIM)Make.defs
-include kernel_test$(DELIM)network$(DELIM)Make.defs
-CSRCS += kernel_test_drv.c
-DEPPATH += --dep-path kernel_test
-VPATH += :kernel_test
-endif
