+++ os/drivers/os_api_test/Make.defs
@@ -1,6 +1,6 @@
-##########################################################################
+###########################################################################
 #
-# Copyright 2020 Samsung Electronics All Rights Reserved.
+# Copyright 2017 Samsung Electronics All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,13 +14,12 @@
 # either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 #
-############################################################################
+###########################################################################
 
-ifeq ($(CONFIG_TC_NET_PBUF),y)
-CSRCS += test_net_pbuf.c
+ifeq ($(CONFIG_DRIVERS_KERNEL_TEST),y)
+include os_api_test$(DELIM)kernel$(DELIM)Make.defs
+include os_api_test$(DELIM)network$(DELIM)Make.defs
+CSRCS += os_api_test_drv.c
+DEPPATH += --dep-path os_api_test
+VPATH += :os_api_test
 endif
-
-# Include network test driver support
-
-DEPPATH += --dep-path kernel_test$(DELIM)network
-VPATH += :kernel_test$(DELIM)network
