+++ os/drivers/os_api_test/kernel/Make.defs
@@ -1,0 +1,52 @@
+##########################################################################
+#
+# Copyright 2020 Samsung Electronics All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+# either express or implied. See the License for the specific
+# language governing permissions and limitations under the License.
+#
+############################################################################
+# Include OS API test drivers
+
+ifeq ($(CONFIG_DRIVERS_KERNEL_TEST),y)
+
+CSRCS += test_task.c
+CSRCS += test_sched.c
+CSRCS += test_clock.c
+CSRCS += test_timer.c
+CSRCS += test_sem.c
+CSRCS += test_group.c
+
+ifneq ($(CONFIG_DISABLE_SIGNALS),y)
+CSRCS += test_signal.c
+endif
+
+ifeq ($(CONFIG_TC_COMPRESS_READ),y)
+CSRCS += test_compress_decompress.c
+endif
+
+ifeq ($(CONFIG_ARMV8M_TRUSTZONE),y)
+CSRCS += test_tz.c
+endif
+
+ifeq ($(CONFIG_EXAMPLES_STACK_PROTECTION),y)
+CSRCS += test_kthread_stack_protection.c
+endif
+
+CFLAGS += -I$(TOPDIR)/kernel
+
+# Include OS API test driver support
+
+DEPPATH += --dep-path os_api_test$(DELIM)kernel
+VPATH += :os_api_test$(DELIM)kernel
+
+endif
