+++ os/drivers/os_api_test/kernel/test_sem.c
@@ -27,7 +27,7 @@
 #include <semaphore.h>
 
 #include <tinyara/sched.h>
-#include <tinyara/kernel_test_drv.h>
+#include <tinyara/os_api_test_drv.h>
 
 #include "clock/clock.h"
