+++ os/drivers/os_api_test/kernel/test_task.c
@@ -30,7 +30,7 @@
 #include <sched.h>
 
 #include <tinyara/sched.h>
-#include <tinyara/kernel_test_drv.h>
+#include <tinyara/os_api_test_drv.h>
 
 #include "sched/sched.h"
 #ifdef CONFIG_SCHED_HAVE_PARENT
